+++ psim/airdrop/masternode/payout_queue.go
@@ -1,0 +1,8 @@
+package masternode
+
+import "time"
+
+type PayoutQueue struct {
+	FirstPayout   time.Time
+	BlockDuration time.Duration
+}