+++ psim/cmd/psim/main.go
@@ -50,6 +50,7 @@ import (
 	_ "gitlab.com/swarmfund/psim/psim/airdrop/20airdrop"
 	_ "gitlab.com/swarmfund/psim/psim/airdrop/earlybird"
 	_ "gitlab.com/swarmfund/psim/psim/airdrop/kycairdrop"
+	_ "gitlab.com/swarmfund/psim/psim/airdrop/masternode"
 	_ "gitlab.com/swarmfund/psim/psim/airdrop/mrefairdrop"
 	_ "gitlab.com/swarmfund/psim/psim/airdrop/telegram"