+++ CHANGELOG.md
@@ -17,3 +17,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     `offchain_currency` and `offchain_blockchain` removed in favor of `network_type`
     `external_system` is now optional, by default `deposit_asset` external system type is used
     withdraw services now have `service_name`/`verifier_service_name` to play nicer with discovery
+
+## Masternode airdrop
+
+### Added
+
+- new airdrop service